// apps/gocho_web/src/component/common/atom/accountInput/type.ts
import { FieldError, UseFormRegisterReturn } from "react-hook-form";

export interface AccountInputProps {
  placeholder: string;
  label: string;
  registerObj: UseFormRegisterReturn;
  errorObj: FieldError | undefined;
  isDirty: boolean | undefined;
  inputType: "text" | "password" | "email";
}

export type ColorStateDef = "error" | "focus" | "default" | "success";